+++ vite.config.ts
@@ -7,6 +7,9 @@ import vue from "@vitejs/plugin-vue";
 // https://vitejs.dev/config/
 export default defineConfig({
     plugins: [vue()],
+    build: {
+        outDir: "./ui/dist/",
+    },
     resolve: {
         alias: {
             "@": fileURLToPath(new URL("./ui", import.meta.url)),